visualizer/requirements.txt
Cython==0.29.27
decorator==5.1.1
freetype-py==2.2.0
imageio==2.16.0
lxml==4.7.1
networkx==2.2
numpy==1.22.2
Pillow==9.0.1
pycollada==0.6
pyglet==1.5.14
PyOpenGL==3.1.5
pyrender==0.1.43
python-dateutil==2.8.2
python-fcl==0.6.1
PyYAML==6.0
scipy==1.8.0
six==1.16.0
trimesh==3.10.0
urdfpy==0.0.22
